# src/main/resources/static/openapi.yaml
openapi: "3.0.2"
info:
  title: Freelancer Web Services API
  version: "0.1"
  description: >
    The Freelancer Web Services API is a RESTful API that allows the registration and retrieval of freelancers.
    It is a simple API that serves as a reference for the development of RESTful APIs.
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /projects:
    get:
      tags:
        - Projects
      description: Returns a paginated list of projects.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: 0-based page number to list
        - in: query
          name: size
          schema:
            type: integer
          required: false
          description: Number of elements per page
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                type: object
                required:
                  - _embedded
                  - _links
                properties:
                  _embedded:
                    type: object
                    required:
                      - projects
                    properties:
                      projects:
                        type: array
                        items:
                          $ref: "#/components/schemas/Project"
                  _links:
                    type: object
                    required:
                      - self
                      - create
                    properties:
                      self:
                        $ref: "#/components/schemas/Link"
                      create:
                        $ref: "#/components/schemas/Link"
                      first:
                        $ref: "#/components/schemas/Link"
                      prev:
                        $ref: "#/components/schemas/Link"
                      next:
                        $ref: "#/components/schemas/Link"
                      last:
                        $ref: "#/components/schemas/Link"
                  page:
                    $ref: "#/components/schemas/Page"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Projects
      description: Creates a new project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "201":
          description: Project was created.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /projects/{id}:
    get:
      tags:
        - Projects
      description: Retrieves a project by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a project already created.
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /projects/{id}/join:
    post:
      tags:
        - Projects
      description: Allows a freelancer to join a project.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of an existing project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                freelancerId:
                  type: string
                  format: uuid
                  description: Identifier of the freelancer that will join the project
                  example: "fa8508ed-8b7b-4be7-b372-ac1094c709b5"
                committedHours:
                  type: integer
                  description: Number of hours per week that the freelancer will commit to the project
                  example: 100
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
components:
  schemas:
...
    Project:
      type: object
      description: Information about a project
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: Unique identifier for the project
          example: "fa8508ed-8b7b-4be7-b372-ac1094c709b5"
        name:
          type: string
          description: Name of the project
          example: "Avengers"
        description:
          type: string
          description: Description of the project
          example: "Earth's mightiest heroes"
        requiredHours:
          type: integer
          description: Required hours for the project
          example: 1000
        _embedded:
          type: object
          description: Embedded resources for the project
          readOnly: true
          properties:
            freelancers:
              type: array
              description: List of freelancers assigned to the project
              items:
                type: object
                properties:
                  id:
                    type: string
                    description: Unique identifier for the freelancer
                    example: "fa8508ed-8b7b-4be7-b372-ac1094c709b5"
                  firstName:
                    type: string
                    description: First name of the freelancer
                    example: "Tony"
                  lastName:
                    type: string
                    description: Last name of the freelancer
                    example: "Stark"
                  allocatedHours:
                    type: integer
                    description: Allocated hours for the freelancer in the project
                    example: 100
                  _links:
                    type: object
                    description: Links for subresources and actions available for the freelancer
                    readOnly: true
                    properties:
                      self:
                        $ref: "#/components/schemas/Link"
                    required:
                      - self
                required:
                  - id
                  - firstName
                  - lastName
                  - allocatedHours
                  - _links
          required:
            - freelancers
        _links:
          type: object
          description: Links for subresources and actions available for the project
          readOnly: true
          properties:
            collection:
              $ref: "#/components/schemas/Link"
            self:
              $ref: "#/components/schemas/Link"
            join:
              $ref: "#/components/schemas/Link"
          required:
            - collection
            - self
      required:
        - name
        - description
        - requiredHours
        - _embedded
        - _links
    Page:
      type: object
      properties:
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 1
        number:
          type: integer
          example: 0
    Link:
      type: object
      description: Link to a resource or action
      properties:
        href:
          type: string
          description: URI of the resource or action
          example: "/collection-name/resource-id"
        method:
          type: string
          description: HTTP method for the link
          example: "GET"
        templated:
          type: boolean
          description: Whether the URI contains template parameters
          example: false
      required:
        - href
        - method
  responses:
    Error:
      description: >
        Error response formatted as per the Internet Standard
        <a href="https://datatracker.ietf.org/doc/html/rfc7807">Problem Details for HTTP APIs (RFC 7807)</a>.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                description: A URI reference that identifies the problem type.
                type: string
                format: uri-reference
                example: /probs/validation-error
              title:
                description: A short, human-readable summary of the problem type.
                type: string
                example: Bad Request
              status:
                description: The HTTP status code generated by the origin server for this occurrence of the problem.
                type: integer
                example: 400
              detail:
                description: A human-readable explanation specific to this occurrence of the problem.
                type: string
                example: Invalid request content.
              instance:
                description: >
                  A URI reference that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.
                type: string
                format: uri-reference
                example: /freelancers
              errors:
                description: Optional array that describes validation errors.
                type: array
                items:
                  type: object
                  required:
                    - detail
                  properties:
                    attribute:
                      description: >
                        Name of the request body's attribute that caused the problem,
                        if not present the whole body should be considered the cause of the problem.
                      type: string
                      example: firstName
                    detail:
                      description: A human-readable description of the error
                      type: string
                      example: must not be blank
            required:
              - type
              - title
              - status
              - instance